feat(types): add ResourceType and ServerIDs type definitions refactor(models, services, utils): update imports to use new types
