+++ packages/ai/mcp-server/src/devopness_mcp_server/lib/services/service_service.py
@@ -6,8 +6,9 @@ from devopness.models import (
 )
 
 from ..devopness_api import devopness, ensure_authenticated
-from ..models import ActionSummary, ServerIDs, ServiceSummary
+from ..models import ActionSummary, ServiceSummary
 from ..response import MCPResponse
+from ..types import ServerIDs
 from ..utils import (
     get_instructions_format_list,
     get_instructions_format_resource,
